# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  handlers.CreateTiffinBillRequest:
    properties:
      amount:
        type: number
      companyID:
        type: string
      date:
        type: string
      employeeID:
        type: string
      month:
        type: integer
      remarks:
        type: string
      year:
        type: integer
    type: object
...
host: localhost:5000
info:
  contact:
    email: support@peoplehub.com
    name: PeopleHub Team
  description: Enterprise HR Management & Payroll System API.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  title: PeopleHub API
  version: "1.1"
paths:
...
  /tiffin-bills:
    get:
      description: Get all tiffin bills for a company, optionally filtered by month/year
      parameters:
      - description: Company ID
        in: query
        name: company_id
        type: string
      - description: Month (1-12)
        in: query
        name: month
        type: integer
      - description: Year
        in: query
        name: year
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: List tiffin bills
      tags:
      - Payroll
    post:
      consumes:
      - application/json
      description: Create a new tiffin bill entry for an employee
      parameters:
      - description: Tiffin bill details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.CreateTiffinBillRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create tiffin bill
      tags:
      - Payroll
  /tiffin-bills/{id}:
    delete:
      description: Soft delete a tiffin bill
      parameters:
      - description: Tiffin Bill ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Delete tiffin bill
      tags:
      - Payroll
    put:
      consumes:
      - application/json
      description: Update an existing tiffin bill
      parameters:
      - description: Tiffin Bill ID
        in: path
        name: id
        required: true
        type: string
      - description: Updated fields
        in: body
        name: request
        schema:
          $ref: '#/definitions/handlers.CreateTiffinBillRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update tiffin bill
      tags:
      - Payroll
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"